Reorganizing schema page specific components
